refactor: common의 TransactionType으로 변경
