+++ app/src/main/java/com/fintern/ourbudgeting/ui/statistics/chart/components/StatisticsTab.kt
@@ -1,6 +1,6 @@
 package com.fintern.ourbudgeting.ui.statistics.chart.components
 
-import com.fintern.ourbudgeting.ui.statistics.chart.model.TransactionType
+import com.fintern.ourbudgeting.ui.common.model.TransactionType
 
 enum class StatisticsTab(
     val type: TransactionType,
