--- app/src/main/java/com/fintern/ourbudgeting/ui/statistics/chart/model/TransactionType.kt
@@ -1,3 +1,0 @@
-package com.fintern.ourbudgeting.ui.statistics.chart.model
-
-enum class TransactionType { EXPENSE, INCOME }
